backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic-settings==2.8.1
redis==5.2.1
openai==1.68.2

